+++ .storybook/manager.js
@@ -1,0 +1,5 @@
+import '@storybook/addon-knobs/register';
+import '@storybook/addon-a11y/register';
+import '@storybook/addon-actions/register';
+import '@storybook/addon-viewport/register';
+import 'storybook-readme/register'; // Currently used in body of story only, not in add-on panel