+++ src/shared-components/alert/style.js
@@ -4,11 +4,26 @@ import { css, keyframes } from '@emotion/core';
 import {
   COLORS,
   MEDIA_QUERIES,
-  BOX_SHADOWS,
   SPACER,
   ANIMATION,
   TYPOGRAPHY_CONSTANTS,
 } from '../../constants';
+
+export const AlertsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  position: fixed;
+  top: ${SPACER.medium};
+  left: 0;
+  right: 0;
+  z-index: 99999;
+
+  ${MEDIA_QUERIES.mdUp} {
+    left: auto;
+    right: ${SPACER.medium};
+  }
+`;
 
 const fadeInDesktop = keyframes`
   from { opacity: 0; transform: translate3d(24px, 0, 0); }
@@ -20,51 +35,39 @@ const fadeInMobile = keyframes`
   to { opacity: 1; transform: translate3d(0, 0px, 0); }
 `;
 
+const defaultAlertStyles = css`
+  background-color: ${COLORS.primary};
+  box-shadow: 0px 8px 24px rgba(51, 46, 84, 0.05);
+`;
+
 const successAlertStyles = css`
-  background-color: ${COLORS.successBackground};
-  border-color: ${COLORS.successBorder};
-  color: ${COLORS.success};
-  fill: ${COLORS.success};
+  background-color: ${COLORS.success};
+  box-shadow: 0px 8px 24px rgba(43, 110, 51, 0.05);
 `;
 
 const errorAlertStyles = css`
-  background-color: ${COLORS.errorBackground};
-  border-color: ${COLORS.errorBorder};
-  color: ${COLORS.error};
-  fill: ${COLORS.error};
-`;
-
-const defaultAlertStyles = css`
-  background-color: ${COLORS.defaultBackground};
-  border-color: ${COLORS.defaultBorder};
-  color: ${COLORS.default};
-  fill: ${COLORS.default};
+  background-color: ${COLORS.error};
+  box-shadow: 0px 8px 24px rgba(189, 32, 15, 0.05);
 `;
 
 export const AlertContainer = styled.div`
-  animation: ${fadeInMobile} ${ANIMATION.defaultTiming} 1;
-  border-width: 1px;
-  border-style: solid;
-  box-shadow: ${BOX_SHADOWS.message};
   cursor: pointer;
-  display: flex;
-  font-size: ${TYPOGRAPHY_CONSTANTS.fontSize.caption};
-  margin: 0 auto ${SPACER.small};
-  opacity: ${props => (props.exiting ? '0' : '1')};
-  padding: 11px 0px;
-  display: flex;
-  align-items: center;
   position: relative;
-  justify-content: flex-start;
-  text-align: left;
-  transform: ${props =>
-    props.exiting ? 'translate3d(0, -24px, 0)' : 'translate3d(0, 0, 0)'};
-  transition: ${ANIMATION.defaultTiming};
+  margin: 0 auto ${SPACER.small};
+  padding: 0;
   width: 327px;
+  border-radius: ${SPACER.small};
+  opacity: ${props => (props.exiting ? '0' : '1')};
+  animation: ${fadeInMobile} ${ANIMATION.defaultTiming} 1;
+  transition: ${ANIMATION.defaultTiming};
+  transform: ${({ exiting }) =>
+    exiting ? 'translate3d(0, -24px, 0)' : 'translate3d(0, 0, 0)'};
 
-  ${props => {
-    switch (props.alertType) {
+  ${({ alertType }) => {
+    switch (alertType) {
       case 'danger':
+        return errorAlertStyles;
+      case 'error':
         return errorAlertStyles;
       case 'success':
         return successAlertStyles;
@@ -73,40 +76,51 @@ export const AlertContainer = styled.div`
     }
   }};
 
-  ${MEDIA_QUERIES.lgUp} {
-    animation: ${fadeInDesktop} ${ANIMATION.defaultTiming} 1;
-    font-size: ${TYPOGRAPHY_CONSTANTS.fontSize.caption};
-    margin-bottom: ${SPACER.medium};
-    transform: ${props =>
-      props.exiting ? 'translate3d(24px, 0, 0)' : 'translate3d(0, 0, 0)'};
-  }
-`;
-
-export const AlertsContainer = styled.div`
-  align-items: flex-end;
-  display: flex;
-  flex-direction: column;
-  left: 0;
-  position: fixed;
-  right: 0;
-  top: ${SPACER.medium};
-  z-index: 99999;
-
   ${MEDIA_QUERIES.mdUp} {
-    left: auto;
-    right: ${SPACER.medium};
-    top: ${SPACER.medium};
+    animation: ${fadeInDesktop} ${ANIMATION.defaultTiming} 1;
+    margin: 0 auto ${SPACER.medium};
+    transform: ${({ exiting }) =>
+      exiting ? 'translate3d(24px, 0, 0)' : 'translate3d(0, 0, 0)'};
   }
 `;
 
-export const AlertContentContainer = styled.div`
+export const MainContainer = styled.div`
   display: flex;
-  padding: 0 ${SPACER.large} 0 ${SPACER.medium};
-  width: 100%;
+  flex-flow: row nowrap;
+  justify-content: flex-start;
+  align-items: flex-start;
+  color: ${COLORS.white};
+  font-size: ${TYPOGRAPHY_CONSTANTS.fontSize.caption};
+  padding: ${SPACER.medium};
 `;
 
-export const alertIconStyles = css`
-  margin: 2.5px ${SPACER.medium} 0px 0px;
-  min-height: ${SPACER.medium};
-  min-width: ${SPACER.medium};
+export const ContentContainer = styled.div`
+  margin: -3px 0 0 ${SPACER.medium};
+  max-height: 48px;
+`;
+
+export const CtaContent = styled.div`
+  padding: ${SPACER.medium};
+
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: space-between;
+  align-items: center;
+  color: ${COLORS.white};
+  font-size: ${TYPOGRAPHY_CONSTANTS.fontSize.caption};
+  border-top: 1px solid rgba(255, 255, 255, 0.1);
+`;
+
+export const IconContainer = styled.div`
+  ${({ hasAvatar }) =>
+    hasAvatar &&
+    css`
+      min-width: 32px;
+    `};
+
+  svg {
+    height: ${SPACER.medium};
+    width: ${SPACER.medium};
+    fill: ${COLORS.white};
+  }
 `;