+++ src/shared-components/alert/test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
+import TestRenderer from 'react-test-renderer';
 import { shallow } from 'enzyme';
 
 import Alert from './index';
@@ -9,45 +9,58 @@ const testAlert = {
   type: 'success',
   duration: 'sticky',
 };
-const SampleContentComponent = () => (<div>Your info is updated!  <a href='http://google.com'> Click here to see changes</a></div>);
+
+const CustomContentComponent = () => (
+  <div>
+    <strong>Your info is updated!</strong>{' '}
+    <a href="http://google.com"> Click here to see changes</a>
+  </div>
+);
+
+const createNodeMock = element => {
+  if (element.type === 'div') {
+    return {
+      innerHTML: '',
+    };
+  }
+  return null;
+};
 
 describe('Alert UI snapshots', () => {
   test('renders success type and text', () => {
-    const component = renderer.create(
+    const component = TestRenderer.create(
       <Alert
         content={testAlert.text}
         type="success"
         duration={testAlert.duration}
-        onExit={() => { }}
-      />
+        onExit={() => {}}
+      />,
+      { createNodeMock },
     );
 
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
-  test('renders danger type and text', () => {
-    const component = renderer.create(
+  test('renders error type and text', () => {
+    const component = TestRenderer.create(
       <Alert
         content={testAlert.text}
-        type="danger"
+        type="error"
         duration={testAlert.duration}
-        onExit={() => { }}
-      />
+        onExit={() => {}}
+      />,
+      { createNodeMock },
     );
 
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
-  test('renders info type and text', () => {
-    const component = renderer.create(
-      <Alert
-        content={testAlert.text}
-        type="info"
-        duration={testAlert.duration}
-        onExit={() => { }}
-      />
+  test('renders default type and text', () => {
+    const component = TestRenderer.create(
+      <Alert content={testAlert.text} onExit={() => {}} />,
+      { createNodeMock },
     );
 
     const tree = component.toJSON();
@@ -55,26 +68,58 @@ describe('Alert UI snapshots', () => {
   });
 
   test('renders custom component passed in content prop', () => {
-    const component = renderer.create(
+    const component = TestRenderer.create(
       <Alert
-        content= {<SampleContentComponent />}
+        content={<CustomContentComponent />}
         type="success"
         duration={testAlert.duration}
-        onExit={() => { }}
-      />
+        onExit={() => {}}
+      />,
+      { createNodeMock },
     );
 
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
-});
 
-test('Alert onExit is triggered on click', () => {
-  jest.useFakeTimers();
-  const spy = jest.fn();
-  const alert = shallow(<Alert onExit={spy} {...testAlert} />);
+  test('Alert onExit is triggered on click', () => {
+    jest.useFakeTimers();
+    const spy = jest.fn();
+    const component = TestRenderer.create(
+      <Alert
+        content={<CustomContentComponent />}
+        type="success"
+        duration={testAlert.duration}
+        onExit={spy}
+      />,
+      { createNodeMock },
+    );
 
-  alert.simulate('click');
-  jest.runAllTimers();
-  expect(spy).toHaveBeenCalled();
+    component.root.findByType('div').props.onClick();
+    jest.runAllTimers();
+    expect(spy).toHaveBeenCalled();
+  });
+
+  test('Alert with CTA', () => {
+    jest.useFakeTimers();
+    const spy = jest.fn();
+    const component = TestRenderer.create(
+      <Alert
+        content={
+          <div>
+            <strong>Whoops!</strong> There was an error updating your address,
+            pleas try again later
+          </div>
+        }
+        type="error"
+        ctaContent="Update Payment Method"
+        onExit={spy}
+      />,
+      { createNodeMock },
+    );
+
+    component.root.findByType('div').props.onClick();
+    jest.runAllTimers();
+    expect(spy).toHaveBeenCalled();
+  });
 });