+++ src/shared-components/button/style.js
@@ -122,9 +122,9 @@ export const baseButtonStyles = ({
 }) => css`
   ${TYPOGRAPHY_STYLE.button};
   appearance: none;
-  border-radius: 0;
+  border-radius: ${SPACER.xsmall};
   border-style: solid;
-  border-width: 2px;
+  border-width: 1px;
   cursor: pointer;
   display: block;
   margin: 0;