+++ src/shared-components/immersiveModal/index.js
@@ -6,7 +6,7 @@ import { RoundButton } from '../button';
 import KEYCODES from '../../constants/keycodes';
 import keyPressMatch from '../../utils/keyPressMatch';
 import OffClickWrapper from '../offClickWrapper';
-import CloseIcon from '../../svgs/icons/close-icon.svg';
+import CrossIcon from '../../svgs/icons/cross-icon.svg';
 import {
   ModalContainer,
   Overlay,
@@ -27,8 +27,9 @@ class ImmersiveModal extends React.Component {
   };
 
   static defaultProps = {
+    header: null,
     canBeClosed: true,
-    onClose: () => {},
+    onClose: () => undefined,
     scrollContainerId: 'modalScrollContainer',
   };
 
@@ -85,6 +86,7 @@ class ImmersiveModal extends React.Component {
       ...rest
     } = this.props;
     return ReactDOM.createPortal(
+      // eslint-disable-next-line react/jsx-props-no-spreading
       <Overlay {...rest}>
         <ModalContainer id={scrollContainerId}>
           <OffClickWrapper onOffClick={this.closeModal}>
@@ -92,7 +94,7 @@ class ImmersiveModal extends React.Component {
               <CloseIconContainer>
                 <RoundButton
                   buttonType="action"
-                  icon={<CloseIcon />}
+                  icon={<CrossIcon />}
                   onClick={onClose}
                 />
               </CloseIconContainer>
@@ -102,7 +104,7 @@ class ImmersiveModal extends React.Component {
           </OffClickWrapper>
         </ModalContainer>
       </Overlay>,
-      this.domNode
+      this.domNode,
     );
   }
 }