+++ src/shared-components/modal/index.js
@@ -4,7 +4,7 @@ import ReactModal from 'react-modal';
 import { Global } from '@emotion/core';
 
 import { RoundButton } from '../button';
-import CloseIcon from '../../svgs/icons/close-icon.svg';
+import CrossIcon from '../../svgs/icons/cross-icon.svg';
 import keyPressMatch from '../../utils/keyPressMatch';
 import KEYCODES from '../../constants/keycodes';
 import {
@@ -27,7 +27,7 @@ class Modal extends React.Component {
   };
 
   static defaultProps = {
-    onClose: () => {},
+    onClose: () => undefined,
     isOpen: false,
     className: '',
     children: '',
@@ -47,7 +47,7 @@ class Modal extends React.Component {
     parentNode: document.querySelector('#reactPortalSection') || document.body,
   };
 
-  componentWillMount() {
+  UNSAFE_componentWillMount() {
     const el = document.querySelector('#reactPortalSection')
       ? '#reactPortalSection'
       : 'body';
@@ -91,7 +91,9 @@ class Modal extends React.Component {
 
   render() {
     const { isVisible } = this.state;
-    const { isOpen, children, className, canBeClosed } = this.props;
+    const {
+      isOpen, children, className, canBeClosed, 
+    } = this.props;
 
     return (
       <React.Fragment>
@@ -122,7 +124,7 @@ class Modal extends React.Component {
                 <ModalCloseIcon>
                   <RoundButton
                     buttonType="action"
-                    icon={<CloseIcon />}
+                    icon={<CrossIcon />}
                     onClick={this.closeModal}
                   />
                 </ModalCloseIcon>