+++ src/shared-components/selectorButton/test.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import {shallow} from 'enzyme';
+import { shallow } from 'enzyme';
 import renderer from 'react-test-renderer';
 
-import AcneFewGlyph from '../../svgs/glyphs/acne-few-glyph.svg';
+import AcneOneGlyph from '../../svgs/glyphs/acne-one-glyph.svg';
 
 import SelectorButton from './index';
 
@@ -10,8 +10,7 @@ describe('<SelectorButton />', () => {
   describe('UI snapshots', () => {
     test('when children is undefined', () => {
       const tree = renderer
-        .create(<SelectorButton checked={false} onClick={() => {
-        }}/>)
+        .create(<SelectorButton checked={false} onClick={() => {}} />)
         .toJSON();
 
       expect(tree).toMatchSnapshot();
@@ -20,8 +19,7 @@ describe('<SelectorButton />', () => {
     test('when children is a node', () => {
       const tree = renderer
         .create(
-          <SelectorButton checked={false} onClick={() => {
-          }}>
+          <SelectorButton checked={false} onClick={() => {}}>
             SelectorButton Text
           </SelectorButton>,
         )
@@ -33,8 +31,7 @@ describe('<SelectorButton />', () => {
     test('when checked type is primary', () => {
       const tree = renderer
         .create(
-          <SelectorButton checked onClick={() => {
-          }} type='primary'>
+          <SelectorButton checked onClick={() => {}} type="primary">
             SelectorButton Text
           </SelectorButton>,
         )
@@ -46,8 +43,7 @@ describe('<SelectorButton />', () => {
     test('when checked type is secondary', () => {
       const tree = renderer
         .create(
-          <SelectorButton checked onClick={() => {
-          }} type='secondary'>
+          <SelectorButton checked onClick={() => {}} type="secondary">
             SelectorButton Text
           </SelectorButton>,
         )
@@ -59,8 +55,11 @@ describe('<SelectorButton />', () => {
     test('when is checkbox', () => {
       const tree = renderer
         .create(
-          <SelectorButton checked={false} onClick={() => {
-          }} selector="checkbox">
+          <SelectorButton
+            checked={false}
+            onClick={() => {}}
+            selector="checkbox"
+          >
             SelectorButton Text
           </SelectorButton>,
         )
@@ -69,12 +68,16 @@ describe('<SelectorButton />', () => {
       expect(tree).toMatchSnapshot();
     });
 
-
     describe('when Icon added', () => {
-      it('hides icon for checkbox with size \'small\'', () => {
+      it("hides icon for checkbox with size 'small'", () => {
         const tree = renderer.create(
-          <SelectorButton checked={false} onClick={() => {
-          }} selector="checkbox" icon={<AcneFewGlyph width={16} height={16}/>} size="small">
+          <SelectorButton
+            checked={false}
+            onClick={() => {}}
+            selector="checkbox"
+            icon={<AcneOneGlyph width={16} height={16} />}
+            size="small"
+          >
             SelectorButton Text
           </SelectorButton>,
         );
@@ -82,10 +85,15 @@ describe('<SelectorButton />', () => {
         expect(tree).toMatchSnapshot();
       });
 
-      it('hides icon for radio button size \'small\'', () => {
+      it("hides icon for radio button size 'small'", () => {
         const tree = renderer.create(
-          <SelectorButton checked={false} onClick={() => {
-          }} selector="radio" icon={<AcneFewGlyph width={16} height={16}/>} size="small">
+          <SelectorButton
+            checked={false}
+            onClick={() => {}}
+            selector="radio"
+            icon={<AcneOneGlyph width={16} height={16} />}
+            size="small"
+          >
             SelectorButton Text
           </SelectorButton>,
         );
@@ -93,10 +101,15 @@ describe('<SelectorButton />', () => {
         expect(tree).toMatchSnapshot();
       });
 
-      it('displays icon for checkbox with size \'large\'', () => {
+      it("displays icon for checkbox with size 'large'", () => {
         const tree = renderer.create(
-          <SelectorButton checked={false} onClick={() => {
-          }} selector="checkbox" icon={<AcneFewGlyph width={16} height={16}/>} size="large">
+          <SelectorButton
+            checked={false}
+            onClick={() => {}}
+            selector="checkbox"
+            icon={<AcneOneGlyph width={16} height={16} />}
+            size="large"
+          >
             SelectorButton Text
           </SelectorButton>,
         );
@@ -104,10 +117,15 @@ describe('<SelectorButton />', () => {
         expect(tree).toMatchSnapshot();
       });
 
-      it('displays icon for radio button with size \'large\'', () => {
+      it("displays icon for radio button with size 'large'", () => {
         const tree = renderer.create(
-          <SelectorButton checked={false} onClick={() => {
-          }} selector="radio" icon={<AcneFewGlyph width={16} height={16}/>} size="large">
+          <SelectorButton
+            checked={false}
+            onClick={() => {}}
+            selector="radio"
+            icon={<AcneOneGlyph width={16} height={16} />}
+            size="large"
+          >
             SelectorButton Text
           </SelectorButton>,
         );
@@ -117,8 +135,13 @@ describe('<SelectorButton />', () => {
 
       it('displays check mark for checked checkbox', () => {
         const tree = renderer.create(
-          <SelectorButton checked onClick={() => {
-          }} selector="checkbox" icon={<AcneFewGlyph width={16} height={16}/>} size="large">
+          <SelectorButton
+            checked
+            onClick={() => {}}
+            selector="checkbox"
+            icon={<AcneOneGlyph width={16} height={16} />}
+            size="large"
+          >
             SelectorButton Text
           </SelectorButton>,
         );
@@ -128,8 +151,13 @@ describe('<SelectorButton />', () => {
 
       it('displays check mark for checked radio button', () => {
         const tree = renderer.create(
-          <SelectorButton checked onClick={() => {
-          }} selector="radio" icon={<AcneFewGlyph width={16} height={16}/>} size="large">
+          <SelectorButton
+            checked
+            onClick={() => {}}
+            selector="radio"
+            icon={<AcneOneGlyph width={16} height={16} />}
+            size="large"
+          >
             SelectorButton Text
           </SelectorButton>,
         );
@@ -142,7 +170,7 @@ describe('<SelectorButton />', () => {
   describe('onClick callback', () => {
     it('is invoked on click', () => {
       const spy = jest.fn();
-      const wrapper = shallow(<SelectorButton checked={false} onClick={spy}/>);
+      const wrapper = shallow(<SelectorButton checked={false} onClick={spy} />);
 
       wrapper.simulate('click');
       expect(spy).toHaveBeenCalled();