+++ src/svgs/deprecated/index.js
@@ -1,0 +1,48 @@
+// @create-index
+
+export { default as AcneFewGlyph } from './acne-few-glyph.svg';
+export { default as AcneManyGlyph } from './acne-many-glyph.svg';
+export { default as AcneSeveralGlyph } from './acne-several-glyph.svg';
+export { default as AcneSomeGlyph } from './acne-some-glyph.svg';
+export { default as AnxiousFilledEmoji } from './anxious-filled-emoji.svg';
+export { default as BodyGlyph } from './body-glyph.svg';
+export { default as ChatBubbleGlyph } from './chat-bubble-glyph.svg';
+export { default as CheckmarkCircleIcon } from './checkmark-circle-icon.svg';
+export { default as ClipboardStackGlyph } from './clipboard-stack-glyph.svg';
+export { default as CloseIcon } from './close-icon.svg';
+export { default as CreditCardIcon } from './credit-card-icon.svg';
+export { default as DeleteIcon } from './delete-icon.svg';
+export { default as DropletIcon } from './droplet-icon.svg';
+export { default as ElatedFilledEmoji } from './elated-filled-emoji.svg';
+export { default as FacebookSquareIcon } from './facebook-square-icon.svg';
+export { default as FirstAidIcon } from './first-aid-icon.svg';
+export { default as FuriousFilledEmoji } from './furious-filled-emoji.svg';
+export {
+  default as GeneralQuestionsGlyph,
+} from './general-questions-glyph.svg';
+export { default as HandWritingGlyph } from './hand-writing-glyph.svg';
+export { default as HeartFilledEmoji } from './heart-filled-emoji.svg';
+export { default as HugFilledEmoji } from './hug-filled-emoji.svg';
+export { default as LogoCIcon } from './logo-c-icon.svg';
+export { default as LogoCondensedIcon } from './logo-condensed-icon.svg';
+export { default as LogoIcon } from './logo-icon.svg';
+export { default as MessageIcon } from './message-icon.svg';
+export { default as NotificationsIcon } from './notifications-icon.svg';
+export {
+  default as OverwhelmedFilledEmoji,
+} from './overwhelmed-filled-emoji.svg';
+export { default as PhotosIcon } from './photos-icon.svg';
+export { default as PreparedFilledEmoji } from './prepared-filled-emoji.svg';
+export { default as QuestionCircleIcon } from './question-circle-icon.svg';
+export { default as SatisfiedFilledEmoji } from './satisfied-filled-emoji.svg';
+export { default as SkinTextureGlyph } from './skin-texture-glyph.svg';
+export { default as SpaceshipGlyph } from './spaceship-glyph.svg';
+export { default as SparkleFilledEmoji } from './sparkle-filled-emoji.svg';
+export { default as SunscreenBottleGlyph } from './sunscreen-bottle-glyph.svg';
+export { default as SunscreenDropGlyph } from './sunscreen-drop-glyph.svg';
+export { default as SunscreenTubGlyph } from './sunscreen-tub-glyph.svg';
+export { default as SupportedEmoji } from './supported-emoji.svg';
+export { default as SupportedFilledEmoji } from './supported-filled-emoji.svg';
+export {
+  default as UnsatisfiedFilledEmoji,
+} from './unsatisfied-filled-emoji.svg';