+++ src/svgs/logos/index.js
@@ -1,0 +1,4 @@
+// @create-index
+
+export { default as Logo } from './logo.svg';
+export { default as Wordmark } from './wordmark.svg';