+++ src/utils/icons/deprecatedList.js
@@ -1,3 +1,81 @@
 module.exports = {
-  LogoCIcon: 'LogoCIcon will be deprecated in the next version. Use LogoCondensedIcon instead.',
+  AcneFewGlyph:
+    'AcneFewIcon will be deprecated in the next version. Use AcneOneGlyph instead.',
+  AcneManyGlyph:
+    'AcneManyGlyph will be deprecated in the next version. Use AcneSevenGlyph instead.',
+  AcneSeveralGlyph:
+    'AcneSeveralGlyph will be deprecated in the next version. Use AcneFiveGlyph instead.',
+  AcneSomeGlyph:
+    'AcneSomeGlyph will be deprecated in the next version. Use AcneThreeGlyph instead.',
+  AnxiousFilledEmoji:
+    'AnxiousFilledEmoji will be deprecated in the next version. Use AnxiousSelectedEmoji instead.',
+  BodyGlyph:
+    'BodyGlyph will be deprecated in the next version. Use ChestGlyph instead.',
+  ChatBubbleGlyph:
+    'ChatBubbleGlyph will be deprecated in the next version. Use MessageGlyph instead.',
+  CheckmarkCircleIcon:
+    'CheckmarkCircleIcon will be deprecated in the next version. Use CircleCheckmarkIcon instead.',
+  ClipboardStackGlyph:
+    'ClipboardStackGlyph will be deprecated in the next version. Use ConsultHistoryGlyph instead.',
+  CloseIcon:
+    'CloseIcon will be deprecated in the next version. Use CrossIcon instead.',
+  CreditCardIcon:
+    'CreditCardIcon will be deprecated in the next version. Use PaymentIcon instead.',
+  DeleteIcon:
+    'DeleteIcon will be deprecated in the next version. Use TrashIcon instead.',
+  DropletIcon:
+    'DropletIcon will be deprecated in the next version. Use DropIcon instead.',
+  ElatedFilledEmoji:
+    'ElatedFilledEmoji will be deprecated in the next version. Use ElatedSelectedEmoji instead.',
+  FacebookSquareIcon:
+    'FacebookSquareIcon will be deprecated in the next version. Use FacebookIcon instead.',
+  FirstAidIcon:
+    'FirstAidIcon will be deprecated in the next version. Use TreatmentIcon instead.',
+  FuriousFilledEmoji:
+    'FuriousFilledEmoji will be deprecated in the next version. Use FuriousSelectedEmoji instead.',
+  GeneralQuestionsGlyph:
+    'GeneralQuestionsGlyph will be deprecated in the next version. Use MessageGlyph instead.',
+  HandWritingGlyph: 'HandWritingGlyph will be deprecated in the next version.',
+  HeartFilledGlyph:
+    'HeartFilledGlyph will be deprecated in the next version. Use HeartSelectedGlyph instead.',
+  HugFilledGlyph:
+    'HugFilledGlyph will be deprecated in the next version. Use HugSelectedGlyph instead.',
+  LogoCIcon:
+    'LogoCIcon will be deprecated in the next version. Use Logo instead.',
+  LogoCondensedIcon:
+    'LogoCondensedIcon will be deprecated in the next version. Use Logo instead.',
+  LogoIcon:
+    'LogoIcon will be deprecated in the next version. Use Wordmark instead.',
+  MessageIcon:
+    'MessageIcon will be deprecated in the next version. Use ChatBubbleIcon instead.',
+  NotificationsIcon:
+    'NotificationsIcon will be deprecated in the next version. Use BellIcon instead.',
+  OverwhelmedFilledEmoji:
+    'OverwhelmedFilledEmoji will be deprecated in the next version. Use OverwhelmedSelectedEmoji instead.',
+  PhotosIcon:
+    'PhotosIcon will be deprecated in the next version. Use ImageIcon instead.',
+  PreparedFilledEmoji:
+    'PreparedFilledEmoji will be deprecated in the next version. Use PreparedSelectedEmoji instead.',
+  QuestionCircleIcon:
+    'QuestionCircleIcon will be deprecated in the next version. Use HelpIcon instead.',
+  SatisfiedFilledEmoji:
+    'SatisfiedFilledEmoji will be deprecated in the next version. Use SatisfiedSelectedEmoji instead.',
+  SkinTextureGlyph:
+    'SkinTextureGlyph will be deprecated in the next version. Use TextureGlyph instead.',
+  SpaceshipGlyph:
+    'SpaceshipGlyph will be deprecated in the next version. Use RocketshipGlyph instead.',
+  SparkleFilledGlyph:
+    'SparkleFilledGlyph will be deprecated in the next version. Use SparkleSelectedGlyph instead.',
+  SunscreenBottleGlyph:
+    'SunscreenBottleGlyph will be deprecated in the next version. Use SunscreenGlyph instead.',
+  SunscreenDropGlyph:
+    'SunscreenDropGlyph will be deprecated in the next version.',
+  SunscreenTubGlyph:
+    'SunscreenTubGlyph will be deprecated in the next version.',
+  SupportedEmoji:
+    'SupportedEmoji will be deprecated in the next version. Use SmileyEmoji instead.',
+  SupportedFilledEmoji:
+    'SupportedFilledEmoji will be deprecated in the next version. Use SmileySelectedEmoji instead.',
+  UnsatisfiedFilledEmoji:
+    'UnsatisfiedFilledEmoji will be deprecated in the next version. Use UnsatisfiedSelectedEmoji instead.',
 };