+++ stories/accordion/index.tsx
@@ -13,79 +13,71 @@ import AccordionReadme from 'docs/accordion';
 import DefaultAccordion from './defaultAccordion';
 import SubcomponentAccordions from './subcomponentAccordions';
 
-const subcomponents = (
-  <React.Fragment>
-    <Typography.Heading
-      css={css`
-        padding: ${SPACER.large} 0 ${SPACER.small};
-      `}
-    >
-      Subcomponents
-    </Typography.Heading>
-    <SubcomponentAccordions />
-  </React.Fragment>
-);
+const stories = storiesOf('Accordion', module);
 
-const accordionWithKnobs = (
-  <React.Fragment>
-    <Typography.Heading
-      css={css`
-        padding: ${SPACER.small} 0 ${SPACER.small};
-      `}
-    >
-      With Knobs
-    </Typography.Heading>
-    <div
-      css={css`
-        width: 400px;
-      `}
-    >
-      <Accordion.Container>
-        <Accordion
-          title={
-            <Accordion.Content>
-              <Accordion.Truncate>
-                {text('Title text', 'Title')}
-              </Accordion.Truncate>
-            </Accordion.Content>
-          }
-          onClick={action('Accordion clicked')}
-          isOpen={boolean('isOpen', false)}
-          noBorder={boolean('noBorder', false)}
-          disabled={boolean('disabled', false)}
-          rightAlignArrow={boolean('rightAlignArrow', false)}
-        >
-          <Accordion.Content>
-            {text('Expanded text', 'Accordion expanded content')}
-          </Accordion.Content>
-        </Accordion>
-      </Accordion.Container>
-    </div>
-  </React.Fragment>
-);
+stories.addDecorator(withKnobs);
 
-storiesOf('Accordion', module)
-  .addDecorator(withKnobs)
-  .add(
-    'Usage',
-    withDocs(AccordionReadme, () => (
-      <React.Fragment>
-        <div
+stories.add(
+  'Usage',
+  withDocs(AccordionReadme, () => (
+    <React.Fragment>
+      <div
+        css={css`
+          text-align: left;
+        `}
+      >
+        <Typography.Heading
           css={css`
-            text-align: left;
+            padding: ${SPACER.small} 0 ${SPACER.small};
           `}
         >
-          <Typography.Heading
-            css={css`
-              padding: ${SPACER.small} 0 ${SPACER.small};
-            `}
-          >
-            Components
-          </Typography.Heading>
-          <DefaultAccordion />
-          {subcomponents}
-          {accordionWithKnobs}
+          Components
+        </Typography.Heading>
+        <DefaultAccordion />
+
+        <Typography.Heading
+          css={css`
+            padding: ${SPACER.large} 0 ${SPACER.small};
+          `}
+        >
+          Subcomponents
+        </Typography.Heading>
+        <SubcomponentAccordions />
+
+        <Typography.Heading
+          css={css`
+            padding: ${SPACER.small} 0 ${SPACER.small};
+          `}
+        >
+          With Knobs
+        </Typography.Heading>
+        <div
+          css={css`
+            width: 400px;
+          `}
+        >
+          <Accordion.Container>
+            <Accordion
+              title={
+                <Accordion.Content>
+                  <Accordion.Truncate>
+                    {text('Title text', 'Title')}
+                  </Accordion.Truncate>
+                </Accordion.Content>
+              }
+              onClick={action('Accordion clicked')}
+              isOpen={boolean('isOpen', false)}
+              noBorder={boolean('noBorder', false)}
+              disabled={boolean('disabled', false)}
+              rightAlignArrow={boolean('rightAlignArrow', false)}
+            >
+              <Accordion.Content>
+                {text('Expanded text', 'Accordion expanded content')}
+              </Accordion.Content>
+            </Accordion>
+          </Accordion.Container>
         </div>
-      </React.Fragment>
-    ))
-  );
+      </div>
+    </React.Fragment>
+  )),
+);