+++ stories/banner/index.js
@@ -1,10 +1,8 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withDocs } from 'storybook-readme';
-import { action } from '@storybook/addon-actions';
-import { withKnobs, select } from '@storybook/addon-knobs';
-
-import BannerReadme from 'docs/banner.md';
+import { withKnobs, select, text } from '@storybook/addon-knobs';
+import BannerReadme from 'docs/banner';
 import { Banner, Typography } from 'src/shared-components';
 
 const stories = storiesOf('Banner', module);
@@ -15,12 +13,34 @@ stories.add(
   'Usage',
   withDocs(BannerReadme, () => (
     <React.Fragment>
+      <Banner
+        content={
+          <React.Fragment>
+            <strong>Default banner:</strong> This is the banner content
+          </React.Fragment>
+        }
+      />
+      <Banner
+        content={
+          <React.Fragment>
+            <strong>Success banner:</strong> This is the banner content
+          </React.Fragment>
+        }
+        type="success"
+      />
+      <Banner
+        content={
+          <React.Fragment>
+            <strong>Error Banner:</strong> This is the banner content
+          </React.Fragment>
+        }
+        type="error"
+      />
       <Typography.Heading>With Knobs</Typography.Heading>
       <Banner
-        content={<div>This is a banner <a href="#">something</a></div>}
-        type={select('type', ['success', 'info', 'danger'], 'info')}
-        onClick={action('clicked')}
+        type={select('type', ['default', 'success', 'error'], 'default')}
+        content={text('content', 'This is the banner content')}
       />
-    </React.Fragment >
-  ))
+    </React.Fragment>
+  )),
 );